requirements.txt
python-telegram-bot[job-queue]==20.7
nest_asyncio


